requirements/base.txt
pytz==2019.3  # https://github.com/stub42/pytz
python-slugify==4.0.0  # https://github.com/un33k/python-slugify
Pillow==6.2.1  # https://github.com/python-pillow/Pillow
rcssmin==1.0.6 --install-option="--without-c-extensions"  # https://github.com/ndparker/rcssmin
argon2-cffi==19.2.0  # https://github.com/hynek/argon2_cffi
celery==4.4.0  # pyup: < 5.0  # https://github.com/celery/celery
django-celery-beat==1.5.0  # https://github.com/celery/django-celery-beat
django-el-pagination==3.1.0
Markdown==2.6.2
wagtailmenus==2.8
django-widget-tweaks==1.4.2
django-bootstrap4
django-bootstrap-datepicker-plus
django-colorful
django-avatar
django-photologue
django-tables2
django-import-export
django-object-tools

# Django
# ------------------------------------------------------------------------------
django==2.2.9  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.41.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.8.1  # https://github.com/django-crispy-forms/django-crispy-forms
django-compressor==2.3  # https://github.com/django-compressor/django-compressor

# Django REST Framework
djangorestframework==3.11.0  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
